fix: make command registration named module
